add working training
Realized that the model if trained using YOLO method would need the dataset to be in different format. So script to convert MOT images to YOLO format.

script to check if the conversion to YOLO was successful, script changes to identify ball in new dataset using the class_id col in MOT dataset

check for initial MOT conversion