Add eval file
Add a file for evaluation of detection and then run byteTrack.

The detection and tracking of players is good, but the ball tracking is not so much. Might need to retrain model and might need to see the inference on a random video to just get the output.

One of our videos